Defactoring search screen, pulling out user fetch to firesbase methods
